logger -> request_logger
