+++ vllm/entrypoints/speech_to_text/realtime/serving.py
@@ -11,7 +11,7 @@ import numpy as np
 from vllm.engine.protocol import EngineClient, StreamingInput
 from vllm.entrypoints.openai.engine.serving import OpenAIServing
 from vllm.entrypoints.openai.models.serving import OpenAIServingModels
-from vllm.entrypoints.serve.utils.logger import RequestLogger
+from vllm.entrypoints.serve.utils.request_logger import RequestLogger
 from vllm.inputs import PromptType
 from vllm.logger import init_logger
 from vllm.model_executor.models.interfaces import SupportsRealtime
